feat: add some plugins
